requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
